pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netai-anomaly"
version = "0.1.0"
description = "Deep learning models for network anomaly detection using perfSONAR telemetry"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Anirudh", email = "anirudh@example.com"},
]
keywords = ["network", "anomaly-detection", "deep-learning", "perfsonar", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
]
gpu = [
    "torch>=2.0.0",
]

[project.scripts]
netai-train = "netai_anomaly.cli:train_cli"
netai-evaluate = "netai_anomaly.cli:evaluate_cli"
netai-serve = "netai_anomaly.cli:serve_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
